+++ src/net/beeroclock/dbeer/activities/WhereBeerActivity.java
@@ -81,10 +81,10 @@ public class WhereBeerActivity extends ListActivity implements LocationListener 
         // Parse raw drink options... may need to push this to a background task...
         String[] raw_drink_options = getResources().getStringArray(R.array.drink_options_raw);
         pinty.drinkNames = new ArrayList<String>();
-        pinty.drinkExternalIds = new ArrayList<Integer>();
+        pinty.drinkExternalIds = new ArrayList<Long>();
         for (String row : raw_drink_options) {
             String[] bits = row.split(";", 2);  // allow drinks to contain ; if they really want.
-            pinty.drinkExternalIds.add(Integer.valueOf(bits[0]));
+            pinty.drinkExternalIds.add(Long.valueOf(bits[0]));
             pinty.drinkNames.add(bits[1]);
         }
     }
